Update regression model evaluations with R2 and RMSE metrics
This commit updates the evaluation metrics for our regression models in the analysis notebook. The R2 and RMSE values for Linear Regression, Lasso, Ridge, and Elastic Net models have been added for better comparison and selection of the optimal model based on performance metrics. Elastic Net showed the highest R2 and lowest RMSE, indicating it as the most effective model among the evaluated ones